\encoding{UTF-8}
\name{Deprecated functions}
\alias{ade4-deprecated}
\title{Deprecated functions in ade4}
\description{
  The functions/data listed below are deprecated:
  
  - \code{between}: replaced by \code{bca}
  - \code{betweencoinertia}: replaced by \code{bca.coinertia}
  - \code{orthogram}: replaced by \code{orthogram} in the \code{adephylo} package
  - \code{within}: replaced by \code{wca}
  - \code{withincoinertia}: replaced by \code{wca.coinertia}
}
