\name{ade4-internal}
\alias{testdiscrimin}
\alias{testertrace}
\alias{testertracenu}
\alias{testertracenubis}
\alias{testinter}
\alias{testprocuste}
\alias{testmantel}
\alias{testertracerlq}
\alias{testamova}
\alias{dudi.type}
\alias{fac2disj}
\alias{RVrandtestCpp}
\alias{RVintrarandtestCpp}
\alias{procusterandtestCpp}
\alias{testinterCpp}
\alias{inerbetweenCpp}
\alias{testmantelCpp}
\alias{betweenvarCpp}
\alias{testdiscriminCpp}
\title{Internal ade4 functions}
\description{
Internal ade4 functions
}

\details{
These are not to be called by the user.
}
\keyword{ internal }
