\encoding{latin1}
\name{ade4-package}
\alias{ade4-package}
\alias{ade4}
\docType{package}
\title{The ade4 package}

\description{This package is developed in the Biometry and Evolutionary Biology Lab (UMR 5558) - University Lyon 1. 
It contains Data Analysis functions to analyse Ecological and Environmental data in the framework of Euclidean Exploratory methods, hence the name ade4.\cr

ade4 is characterized by (1) the implementation of graphical and statistical functions, (2) the availability of numerical data, (3) the redaction of technical and thematic documentation and (4) the inclusion of bibliographic references. \cr

To cite ade4, please use \code{citation("ade4")}.
 }

\author{Daniel Chessel, Anne B Dufour, Stephane Dray with the contributions from J.R. Lobry, S. Ollier, S. Pavoine and J. Thioulouse}
\references{
  See ade4 website: \url{http://pbil.univ-lyon1.fr/ADE-4/}
}
\keyword{manip}
\keyword{multivariate}
\seealso{\code{ade4TkGUI}, \code{adegenet}, \code{adehabitat}}