\name{adegraphicsLoaded}
\alias{adegraphicsLoaded}
\title{Utility function to test if the adegraphics package is loaded}
\description{This function check if the adegraphics package is loaded, before loading ade4. Mainly used to run examples using either ade4 or adegraphics function.}
\usage{adegraphicsLoaded()}
\value{A logical}
\author{Stéphane Dray (\email{stephane.dray@univ-lyon1.fr})}
\keyword{internal}
