\name{kplot}
\alias{kplot}
\title{Generic Function for Multiple Graphs in a K-tables Analysis}
\description{
	Methods for \code{foucart}, \code{mcoa}, \code{mfa}, \code{pta}, \code{sepan}, \code{sepan.coa} and \code{statis}
}
\usage{
kplot(object, ...)
}
\arguments{
  \item{object}{an object used to select a method}
  \item{\dots}{further arguments passed to or from other methods}
}
\examples{
methods(plot)
methods(scatter)
methods(kplot)
}
\keyword{multivariate}
\keyword{hplot}
