\name{multiblock}
\alias{summary.multiblock}
\alias{print.multiblock}

\title{Display and summarize multiblock objects}
\description{Generic methods print and summary for mulitblock objects}
\usage{
\method{summary}{multiblock}(object, ...)
\method{print}{multiblock}(x, ...)
}

\arguments{
  \item{object}{an object of class multiblock created by \code{\link{mbpls}}
    or \code{\link{mbpcaiv}}}
\item{x}{an object of class multiblock created by \code{\link{mbpls}}
    or \code{\link{mbpcaiv}}}
  \item{\dots}{other arguments to be passed to methods}
}

\references{Bougeard, S. and Dray S. (2018) Supervised Multiblock Analysis in R with the ade4 Package. \emph{Journal of Statistical Software}, \bold{86} (1), 1-17. \url{http://doi.org/10.18637/jss.v086.i01}}

\author{Stéphanie Bougeard (\email{stephanie.bougeard@anses.fr}) and Stéphane Dray (\email{stephane.dray@univ-lyon1.fr})}

\seealso{
\code{\link{mbpls}}, \code{\link{mbpcaiv}}
}

\keyword{multivariate}

