\encoding{UTF-8}
\name{Deprecated functions}
\alias{ade4-deprecated}
\title{Deprecated functions in ade4}
\description{
  The functions/data listed below are deprecated.
  The R code of the deprecated functions are stored for memory in the file \code{ade4-deprecated.R}.
  
  - \code{between}: replaced by \code{bca} \cr
  - \code{betweencoinertia}: replaced by \code{bca.coinertia} \cr
  - \code{orthogram}: replaced by \code{orthogram} in the \code{adephylo} package \cr
  - \code{within}: replaced by \code{wca} \cr
  - \code{withincoinertia}: replaced by \code{wca.coinertia} \cr
}
