\encoding{UTF-8}
\name{Deprecated functions}
\alias{ade4-deprecated}
\title{Deprecated functions in ade4}
\description{
  The functions/data/methods listed below are deprecated.
  The R code of the deprecated functions are stored for memory in the file \code{ade4-deprecated.R}.
  
  - \code{between}: replaced by \code{bca} \cr
  - \code{betweencoinertia}: replaced by \code{bca.coinertia} \cr
  - \code{char2genet}: replaced by \code{df2genind} and \code{genind2genpop} in the \code{adegenet} package \cr
  - \code{count2genet}: replaced by \code{df2genind} and \code{genind2genpop} in the \code{adegenet} package \cr
  - \code{dist.genet}: replaced by \code{dist.genpop} in the \code{adegenet} package \cr
  - \code{EH}: replaced by \code{EH} in the \code{adiv} package \cr
  - \code{freq2genet}: replaced by \code{df2genind} and \code{genind2genpop} in the \code{adegenet} package \cr
  - \code{fuzzygenet}: replaced by \code{df2genind} in the \code{adegenet} package \cr
  - \code{multispati}: replaced by \code{multispati} in the \code{adespatial} package \cr
  - \code{multispati.rtest} \cr
  - \code{multispati.randtest} \cr
  - \code{optimEH}: replaced by \code{optimEH} in the \code{adiv} package \cr
  - \code{orisaved}: replaced by \code{orisaved} in the \code{adiv} package \cr
  - \code{orthogram}: replaced by \code{orthogram} in the \code{adephylo} package \cr
  - \code{plot.multispati}: replaced by \code{plot.multispati} in the \code{adespatial} package \cr
  - \code{print.multispati}: replaced by \code{print.multispati} in the \code{adespatial} package \cr
  - \code{randEH}: replaced by \code{randEH} in the \code{adiv} package \cr
  - \code{summary.multispati}: replaced by \code{summary.multispati} in the \code{adespatial} package \cr
  - \code{within}: replaced by \code{wca} \cr
  - \code{withincoinertia}: replaced by \code{wca.coinertia} \cr
}
