\name{ade4-package}
\alias{ade4-package}
\alias{ade4}
\docType{package}
\title{The ade4 package}

\description{This package is developed in the Biometry and Evolutionary Biology Lab (UMR CNRS 5558) - University Lyon 1. 
It contains Data Analysis functions to analyse Ecological and Environmental data in the framework of Euclidean Exploratory methods, hence the name ade4.\cr

ade4 is characterized by (1) the implementation of graphical and statistical functions, (2) the availability of numerical data, (3) the redaction of technical and thematic documentation and (4) the inclusion of bibliographic references. \cr

To cite ade4, please use \code{citation("ade4")}.
 }

\author{Stéphane Dray, Anne-Béatrice Dufour, and Jean Thioulouse. Contributions from Daniel Borcard, Stéphanie Bougeard, Thibaut Jombart, Pierre Legendre, Jean R. Lobry, Sébastien Ollier, Sandrine Pavoine and Aurélie Siberchicot. Based on earlier work by Daniel Chessel.}

\references{
Dray S and Dufour A (2007). “The ade4 Package: Implementing the Duality Diagram for Ecologists.”
_Journal of Statistical Software_, *22*(4), pp. 1-20. doi: 10.18637/jss.v022.i04 (URL:
http://doi.org/10.18637/jss.v022.i04).

See ade4 website: \url{http://pbil.univ-lyon1.fr/ADE-4/}
}
\keyword{manip}
\keyword{multivariate}
\seealso{\code{ade4TkGUI}, \code{adegenet}, \code{adehabitat}, \code{adegraphics}}
