\name{adegraphicsLoaded}
\alias{adegraphicsLoaded}
\title{Utility function to test if the package adegraphics is loaded}
\description{This function check if the package adegraphics is loaded. Mainly used to run examples using either ade4 or adegraphics function
}
\usage{
adegraphicsLoaded()
}

\value{A logical}

\author{Stéphane Dray (\email{stephane.dray@univ-lyon1.fr})}

\keyword{ internal }
