\name{multiblock}
\alias{summary.multiblock}
\alias{print.multiblock}

\title{Display and summarize multiblock objects}
\description{Generic methods print and summary for mulitblock objects}
\usage{
\method{summary}{multiblock}(object, ...)
\method{print}{multiblock}(x, ...)
}

\arguments{
  \item{object}{an object of class multiblock created by \code{\link{mbpls}}
    or \code{\link{mbpcaiv}}}
\item{x}{an object of class multiblock created by \code{\link{mbpls}}
    or \code{\link{mbpcaiv}}}
  \item{\dots}{other arguments to be passed to methods}
}

\author{St\'ephanie Bougeard (\email{stephanie.bougeard@anses.fr}) and St\'ephane Dray (\email{stephane.dray@univ-lyon1.fr})}

\seealso{
\code{\link{mbpls}}, \code{\link{mbpcaiv}}
}

\keyword{multivariate}

