\name{randtest-internal}
\alias{testdiscrimin}
\alias{testertrace}
\alias{testertracenu}
\alias{testertracenubis}
\alias{testinter}
\alias{testprocuste}
\alias{testmantel}
\alias{testertracerlq}
\alias{testamova}
\title{Internal Permutation Tests functions (in C).}
\description{
Internal Permutation Tests functions (in C)
}
\usage{
}
\details{
These are not to be called by the user.
}
\keyword{ internal }
